requirements.txt
streamlit
pandas
joblib
scikit-learn
altair
google-generativeai>=0.7.0
streamlit-mic-recorder
gTTS
SpeechRecognition
pydub